+++ packages/sdk/src/cross-chain-messenger.ts
@@ -5,13 +5,33 @@ import {
   TransactionReceipt,
   TransactionResponse,
   TransactionRequest,
+  Log,
 } from '@ethersproject/abstract-provider'
 import { Signer } from '@ethersproject/abstract-signer'
-import { ethers, BigNumber, Overrides, CallOverrides } from 'ethers'
-import { sleep, remove0x } from '@eth-optimism/core-utils'
-import { predeploys } from '@eth-optimism/contracts'
+import {
+  ethers,
+  BigNumber,
+  Overrides,
+  CallOverrides,
+  PayableOverrides,
+} from 'ethers'
+import {
+  sleep,
+  remove0x,
+  toHexString,
+  toRpcHexString,
+  hashWithdrawal,
+  encodeCrossDomainMessageV0,
+  hashCrossDomainMessage,
+  L2OutputOracleParameters,
+  BedrockOutputData,
+  BedrockCrossChainMessageProof,
+} from '@eth-optimism/core-utils'
+import { getContractInterface, predeploys } from '@eth-optimism/contracts'
+import * as rlp from 'rlp'
 
 import {
+  CoreCrossChainMessage,
   ICrossChainMessenger,
   OEContracts,
   OEContractsLike,
@@ -42,10 +62,8 @@ import {
   DeepPartial,
   getAllOEContracts,
   getBridgeAdapters,
-  hashCrossChainMessage,
   makeMerkleTreeProof,
   makeStateTrieProof,
-  encodeCrossChainMessage,
   DEPOSIT_CONFIRMATION_BLOCKS,
   CHAIN_BLOCK_TIMES,
 } from './utils'
@@ -59,6 +77,9 @@ export class CrossChainMessenger implements ICrossChainMessenger {
   public bridges: BridgeAdapters
   public depositConfirmationBlocks: number
   public l1BlockTimeSeconds: number
+  public bedrock: boolean
+
+  private _l2OutputOracleParameters: L2OutputOracleParameters
 
   /**
    * Creates a new CrossChainProvider instance.
@@ -72,6 +93,7 @@ export class CrossChainMessenger implements ICrossChainMessenger {
    * @param opts.l1BlockTimeSeconds Optional estimated block time in seconds for the L1 chain.
    * @param opts.contracts Optional contract address overrides.
    * @param opts.bridges Optional bridge address list.
+   * @param opts.bedrock Whether or not to enable Bedrock compatibility.
    */
   constructor(opts: {
     l1SignerOrProvider: SignerOrProviderLike
@@ -82,7 +104,9 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     l1BlockTimeSeconds?: NumberLike
     contracts?: DeepPartial<OEContractsLike>
     bridges?: BridgeAdapterData
+    bedrock?: boolean
   }) {
+    this.bedrock = opts.bedrock ?? false
     this.l1SignerOrProvider = toSignerOrProvider(opts.l1SignerOrProvider)
     this.l2SignerOrProvider = toSignerOrProvider(opts.l2SignerOrProvider)
 
@@ -151,6 +175,26 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     }
   }
 
+  public async getL2OutputOracleParameters(): Promise<L2OutputOracleParameters> {
+    if (this._l2OutputOracleParameters) {
+      return this._l2OutputOracleParameters
+    }
+
+    this._l2OutputOracleParameters = {
+      submissionInterval: (
+        await this.contracts.l1.L2OutputOracle.SUBMISSION_INTERVAL()
+      ).toNumber(),
+      startingBlockNumber: (
+        await this.contracts.l1.L2OutputOracle.STARTING_BLOCK_NUMBER()
+      ).toNumber(),
+      l2BlockTime: (
+        await this.contracts.l1.L2OutputOracle.L2_BLOCK_TIME()
+      ).toNumber(),
+    }
+
+    return this._l2OutputOracleParameters
+  }
+
   public async getMessagesByTransaction(
     transaction: TransactionLike,
     opts: {
@@ -203,6 +247,19 @@ export class CrossChainMessenger implements ICrossChainMessenger {
         return parsed.name === 'SentMessage'
       })
       .map((log) => {
+        // Try to pull out the value field, but only if the very next log is a SentMessageExtraData
+        // event which was introduced in the Bedrock upgrade.
+        let value = ethers.BigNumber.from(0)
+        if (receipt.logs.length > log.logIndex + 1) {
+          const next = receipt.logs[log.logIndex + 1]
+          if (next.address === messenger.address) {
+            const nextParsed = messenger.interface.parseLog(next)
+            if (nextParsed.name === 'SentMessageExtension1') {
+              value = nextParsed.args.value
+            }
+          }
+        }
+
         // Convert each SentMessage log into a message object
         const parsed = messenger.interface.parseLog(log)
         return {
@@ -211,7 +268,8 @@ export class CrossChainMessenger implements ICrossChainMessenger {
           sender: parsed.args.sender,
           message: parsed.args.message,
           messageNonce: parsed.args.messageNonce,
-          gasLimit: parsed.args.gasLimit,
+          value,
+          minGasLimit: parsed.args.gasLimit,
           logIndex: log.logIndex,
           blockNumber: log.blockNumber,
           transactionHash: log.transactionHash,
@@ -373,20 +431,32 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       }
     } else {
       if (receipt === null) {
-        const stateRoot = await this.getMessageStateRoot(resolved)
-        if (stateRoot === null) {
-          return MessageStatus.STATE_ROOT_NOT_PUBLISHED
-        } else {
-          const challengePeriod = await this.getChallengePeriodSeconds()
-          const targetBlock = await this.l1Provider.getBlock(
-            stateRoot.batch.blockNumber
-          )
-          const latestBlock = await this.l1Provider.getBlock('latest')
-          if (targetBlock.timestamp + challengePeriod > latestBlock.timestamp) {
-            return MessageStatus.IN_CHALLENGE_PERIOD
-          } else {
-            return MessageStatus.READY_FOR_RELAY
+        let timestamp: number
+        if (this.bedrock) {
+          const output = await this.getMessageBedrockOutput(resolved)
+          if (output === null) {
+            return MessageStatus.STATE_ROOT_NOT_PUBLISHED
           }
+
+          timestamp = output.l1Timestamp
+        } else {
+          const stateRoot = await this.getMessageStateRoot(resolved)
+          if (stateRoot === null) {
+            return MessageStatus.STATE_ROOT_NOT_PUBLISHED
+          }
+
+          const bn = stateRoot.batch.blockNumber
+          const block = await this.l1Provider.getBlock(bn)
+          timestamp = block.timestamp
+        }
+
+        const challengePeriod = await this.getChallengePeriodSeconds()
+        const latestBlock = await this.l1Provider.getBlock('latest')
+
+        if (timestamp + challengePeriod > latestBlock.timestamp) {
+          return MessageStatus.IN_CHALLENGE_PERIOD
+        } else {
+          return MessageStatus.READY_FOR_RELAY
         }
       } else {
         if (receipt.receiptStatus === MessageReceiptStatus.RELAYED_SUCCEEDED) {
@@ -402,7 +472,14 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     message: MessageLike
   ): Promise<MessageReceipt> {
     const resolved = await this.toCrossChainMessage(message)
-    const messageHash = hashCrossChainMessage(resolved)
+    const messageHash = hashCrossDomainMessage(
+      resolved.messageNonce,
+      resolved.sender,
+      resolved.target,
+      resolved.value,
+      resolved.minGasLimit,
+      resolved.message
+    )
 
     // Here we want the messenger that will receive the message, not the one that sent it.
     const messenger =
@@ -642,9 +719,54 @@ export class CrossChainMessenger implements ICrossChainMessenger {
   }
 
   public async getChallengePeriodSeconds(): Promise<number> {
-    const challengePeriod =
-      await this.contracts.l1.StateCommitmentChain.FRAUD_PROOF_WINDOW()
+    const challengePeriod = this.bedrock
+      ? await this.contracts.l1.OptimismPortal.FINALIZATION_PERIOD_SECONDS()
+      : await this.contracts.l1.StateCommitmentChain.FRAUD_PROOF_WINDOW()
     return challengePeriod.toNumber()
+  }
+
+  public async getMessageBedrockOutput(
+    message: MessageLike
+  ): Promise<BedrockOutputData | null> {
+    const resolved = await this.toCrossChainMessage(message)
+
+    // Outputs are only a thing for L2 to L1 messages.
+    if (resolved.direction === MessageDirection.L1_TO_L2) {
+      throw new Error(`cannot get a state root for an L1 to L2 message`)
+    }
+
+    const l2OutputOracleParameters = await this.getL2OutputOracleParameters()
+
+    // TODO: better way to do this
+    let number =
+      resolved.blockNumber - l2OutputOracleParameters.startingBlockNumber
+    while (number % l2OutputOracleParameters.submissionInterval !== 0) {
+      number++
+    }
+
+    // TODO: Handle old messages from before Bedrock upgrade.
+    const events = await this.contracts.l1.L2OutputOracle.queryFilter(
+      this.contracts.l1.L2OutputOracle.filters.OutputProposed(
+        undefined,
+        undefined,
+        number
+      )
+    )
+
+    if (events.length === 0) {
+      return null
+    }
+
+    // Should not happen
+    if (events.length > 1) {
+      throw new Error(`multiple output roots found for message`)
+    }
+
+    return {
+      outputRoot: events[0].args.l2Output,
+      l1Timestamp: events[0].args.l1Timestamp.toNumber(),
+      l2BlockNumber: events[0].args.l2BlockNumber.toNumber(),
+    }
   }
 
   public async getMessageStateRoot(
@@ -826,8 +948,12 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     // https://docs.soliditylang.org/en/v0.8.4/internals/layout_in_storage.html#mappings-and-dynamic-arrays
     const messageSlot = ethers.utils.keccak256(
       ethers.utils.keccak256(
-        encodeCrossChainMessage(resolved) +
-          remove0x(this.contracts.l2.L2CrossDomainMessenger.address)
+        encodeCrossDomainMessageV0(
+          resolved.target,
+          resolved.sender,
+          resolved.message,
+          resolved.messageNonce
+        ) + remove0x(this.contracts.l2.L2CrossDomainMessenger.address)
       ) + '00'.repeat(32)
     )
 
@@ -848,9 +974,142 @@ export class CrossChainMessenger implements ICrossChainMessenger {
           stateRoot.stateRootIndexInBatch
         ),
       },
-      stateTrieWitness: stateTrieProof.accountProof,
-      storageTrieWitness: stateTrieProof.storageProof,
+      stateTrieWitness: toHexString(rlp.encode(stateTrieProof.accountProof)),
+      storageTrieWitness: toHexString(rlp.encode(stateTrieProof.storageProof)),
     }
+  }
+
+  public async getBedrockMessageProof(
+    message: MessageLike
+  ): Promise<
+    [BedrockCrossChainMessageProof, BedrockOutputData, CoreCrossChainMessage]
+  > {
+    const resolved = await this.toCrossChainMessage(message)
+    if (resolved.direction === MessageDirection.L1_TO_L2) {
+      throw new Error(`can only generate proofs for L2 to L1 messages`)
+    }
+
+    const output = await this.getMessageBedrockOutput(resolved)
+    if (output === null) {
+      throw new Error(`state root for message not yet published`)
+    }
+
+    const receipt = await this.l2Provider.getTransactionReceipt(
+      resolved.transactionHash
+    )
+
+    interface WithdrawalEntry {
+      withdrawalInitiated: any
+      withdrawalInitiatedExtension1: any
+    }
+
+    // Handle multiple withdrawals in the same tx and be backwards
+    // compatible without WithdrawalInitiatedExtension1
+    const logs: Partial<{ number: WithdrawalEntry }> = {}
+    for (const [i, log] of Object.entries(receipt.logs)) {
+      if (log.address === predeploys.OVM_L2ToL1MessagePasser) {
+        const decoded =
+          this.contracts.l2.L2ToL1MessagePasser.interface.parseLog(log)
+        // Find the withdrawal initiated events
+        if (decoded.name === 'WithdrawalInitiated') {
+          logs[log.logIndex] = {
+            withdrawalInitiated: decoded.args,
+            withdrawalInitiatedExtension1: null,
+          }
+          if (receipt.logs[i + 1]) {
+            const next =
+              this.contracts.l2.L2ToL1MessagePasser.interface.parseLog(
+                receipt.logs[i + 1]
+              )
+            if (next.name === 'WithdrawalInitiatedExtension1') {
+              logs[log.logIndex].withdrawalInitiatedExtension1 = next.args
+            }
+          }
+        }
+      }
+    }
+
+    // TODO(tynes): be able to handle transactions that do multiple withdrawals
+    // in a single transaction. Right now just go for the first one.
+    const withdrawal = Object.values(logs)[0]
+    if (!withdrawal) {
+      throw new Error(
+        `Cannot find withdrawal logs for ${resolved.transactionHash}`
+      )
+    }
+
+    const withdrawalHash = hashWithdrawal(
+      withdrawal.withdrawalInitiated.nonce,
+      withdrawal.withdrawalInitiated.sender,
+      withdrawal.withdrawalInitiated.target,
+      withdrawal.withdrawalInitiated.value,
+      withdrawal.withdrawalInitiated.gasLimit,
+      withdrawal.withdrawalInitiated.data
+    )
+
+    // Sanity check
+    if (withdrawal.withdrawalInitiatedExtension1) {
+      if (withdrawal.withdrawalInitiatedExtension1.hash !== withdrawalHash) {
+        throw new Error(`Mismatched withdrawal hashes`)
+      }
+    }
+
+    // TODO: turn into util
+    const preimage = ethers.utils.defaultAbiCoder.encode(
+      ['bytes32', 'uint256'],
+      [withdrawalHash, ethers.constants.HashZero]
+    )
+    const isMessageSent =
+      await this.contracts.l2.L2ToL1MessagePasser.sentMessages(withdrawalHash)
+
+    if (!isMessageSent) {
+      throw new Error(`Withdrawal not initiated on L2`)
+    }
+
+    const messageSlot = ethers.utils.keccak256(preimage)
+
+    const stateTrieProof = await makeStateTrieProof(
+      this.l2Provider as ethers.providers.JsonRpcProvider,
+      output.l2BlockNumber,
+      this.contracts.l2.OVM_L2ToL1MessagePasser.address,
+      messageSlot
+    )
+
+    // Sanity check that the value is set to 1 in the state
+    if (!stateTrieProof.storageValue.eq(1)) {
+      throw new Error(`Withdrawal hash ${withdrawalHash} is not set in state`)
+    }
+
+    const block = await (
+      this.l2Provider as ethers.providers.JsonRpcProvider
+    ).send('eth_getBlockByNumber', [
+      toRpcHexString(output.l2BlockNumber),
+      false,
+    ])
+
+    return [
+      {
+        outputRootProof: {
+          // TODO: Handle multiple versions in the future
+          version: ethers.constants.HashZero,
+          stateRoot: block.stateRoot,
+          withdrawerStorageRoot: stateTrieProof.storageRoot,
+          latestBlockhash: block.hash,
+        },
+        withdrawalProof: ethers.utils.RLP.encode(stateTrieProof.storageProof),
+        // withdrawalProof: toHexString(rlp.encode(stateTrieProof.storageProof)),
+      },
+      output,
+      // TODO(tynes): use better type, typechain?
+      {
+        messageNonce: withdrawal.withdrawalInitiated.nonce,
+        sender: withdrawal.withdrawalInitiated.sender,
+        target: withdrawal.withdrawalInitiated.target,
+        value: withdrawal.withdrawalInitiated.value,
+        minGasLimit: withdrawal.withdrawalInitiated.gasLimit,
+        message: withdrawal.withdrawalInitiated.data,
+      },
+    ]
   }
 
   public async sendMessage(
@@ -890,7 +1149,7 @@ export class CrossChainMessenger implements ICrossChainMessenger {
     message: MessageLike,
     opts?: {
       signer?: Signer
-      overrides?: Overrides
+      overrides?: PayableOverrides
     }
   ): Promise<TransactionResponse> {
     return (opts?.signer || this.l1Signer).sendTransaction(
@@ -1033,21 +1292,36 @@ export class CrossChainMessenger implements ICrossChainMessenger {
         throw new Error(`cannot resend L2 to L1 message`)
       }
 
-      return this.contracts.l1.L1CrossDomainMessenger.populateTransaction.replayMessage(
-        resolved.target,
-        resolved.sender,
-        resolved.message,
-        resolved.messageNonce,
-        resolved.gasLimit,
-        messageGasLimit,
-        opts?.overrides || {}
-      )
+      if (this.bedrock) {
+        return this.populateTransaction.finalizeMessage(resolved, {
+          ...(opts || {}),
+          overrides: {
+            ...opts?.overrides,
+            gasLimit: messageGasLimit,
+          },
+        })
+      } else {
+        const legacyL1XDM = new ethers.Contract(
+          this.contracts.l1.L1CrossDomainMessenger.address,
+          getContractInterface('L1CrossDomainMessenger'),
+          this.l1SignerOrProvider
+        )
+        return legacyL1XDM.populateTransaction.replayMessage(
+          resolved.target,
+          resolved.sender,
+          resolved.message,
+          resolved.messageNonce,
+          resolved.minGasLimit,
+          messageGasLimit,
+          opts?.overrides || {}
+        )
+      }
     },
 
     finalizeMessage: async (
       message: MessageLike,
       opts?: {
-        overrides?: Overrides
+        overrides?: PayableOverrides
       }
     ): Promise<TransactionRequest> => {
       const resolved = await this.toCrossChainMessage(message)
@@ -1055,15 +1329,58 @@ export class CrossChainMessenger implements ICrossChainMessenger {
         throw new Error(`cannot finalize L1 to L2 message`)
       }
 
-      const proof = await this.getMessageProof(resolved)
-      return this.contracts.l1.L1CrossDomainMessenger.populateTransaction.relayMessage(
-        resolved.target,
-        resolved.sender,
-        resolved.message,
-        resolved.messageNonce,
-        proof,
-        opts?.overrides || {}
-      )
+      if (this.bedrock) {
+        const [proof, output, withdrawalTx] = await this.getBedrockMessageProof(
+          message
+        )
+        if (!opts) {
+          opts = {}
+        }
+        if (!opts.overrides) {
+          opts.overrides = {}
+        }
+        if (!opts.overrides.value) {
+          opts.overrides.value = withdrawalTx.value
+        }
+
+        return this.contracts.l1.OptimismPortal.populateTransaction.finalizeWithdrawalTransaction(
+          [
+            withdrawalTx.messageNonce,
+            withdrawalTx.sender,
+            withdrawalTx.target,
+            withdrawalTx.value,
+            withdrawalTx.minGasLimit,
+            withdrawalTx.message,
+          ],
+          output.l2BlockNumber,
+          [
+            proof.outputRootProof.version,
+            proof.outputRootProof.stateRoot,
+            proof.outputRootProof.withdrawerStorageRoot,
+            proof.outputRootProof.latestBlockhash,
+          ],
+          proof.withdrawalProof,
+          opts.overrides
+        )
+      } else {
+        // L1CrossDomainMessenger relayMessage is the only method that isn't fully backwards
+        // compatible, so we need to use the legacy interface. When we fully upgrade to Bedrock we
+        // should be able to remove this code.
+        const proof = await this.getMessageProof(resolved)
+        const legacyL1XDM = new ethers.Contract(
+          this.contracts.l1.L1CrossDomainMessenger.address,
+          getContractInterface('L1CrossDomainMessenger'),
+          this.l1SignerOrProvider
+        )
+        return legacyL1XDM.populateTransaction.relayMessage(
+          resolved.target,
+          resolved.sender,
+          resolved.message,
+          resolved.messageNonce,
+          proof,
+          opts?.overrides || {}
+        )
+      }
     },
 
     depositETH: async (
@@ -1071,9 +1388,25 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       opts?: {
         recipient?: AddressLike
         l2GasLimit?: NumberLike
-        overrides?: Overrides
+        overrides?: PayableOverrides
       }
     ): Promise<TransactionRequest> => {
+      if (this.bedrock) {
+        const value = BigNumber.from(opts?.overrides?.value ?? 0)
+        if (!value.eq(0) && !value.eq(amount)) {
+          throw new Error(`amount and value mismatch`)
+        }
+        if (!opts) {
+          opts = {}
+        }
+        if (!opts.overrides) {
+          opts.overrides = {}
+        }
+        if (!opts.overrides.value) {
+          opts.overrides.value = value
+        }
+      }
+
       return this.bridges.ETH.populateTransaction.deposit(
         ethers.constants.AddressZero,
         predeploys.OVM_ETH,
@@ -1086,9 +1419,24 @@ export class CrossChainMessenger implements ICrossChainMessenger {
       amount: NumberLike,
       opts?: {
         recipient?: AddressLike
-        overrides?: Overrides
+        overrides?: PayableOverrides
       }
     ): Promise<TransactionRequest> => {
+      const value = BigNumber.from(opts?.overrides?.value ?? 0)
+      if (this.bedrock) {
+        if (!value.eq(0) && !value.eq(amount)) {
+          throw new Error(`amount and value mismatch`)
+        }
+        if (!opts) {
+          opts = {}
+        }
+        if (!opts.overrides) {
+          opts.overrides = {}
+        }
+        if (!opts.overrides.value) {
+          opts.overrides.value = value
+        }
+      }
       return this.bridges.ETH.populateTransaction.withdraw(
         ethers.constants.AddressZero,
         predeploys.OVM_ETH,